pyproject.toml
[tool.black]
line-length = 99
target-version = ["py39", "py310", "py311", "py312"]
exclude = '''
/(
    \.git|.venv
)/
'''

[tool.isort]
profile = "black"
float_to_top = true
skip_glob = ".venv"

[project]
name = "sqlalchemy-schema"
version = "0.4.2"
description = "Convert SQLAlchemy models to JSON Schema definitions"
repository = "https://github.com/expobrain/sqlalchemy-to-json-schema"
authors = [
    { name = "Daniele Esposti", email = "daniele.esposti@gmail.com" },
    { name = "Jacob Reimers", email = "pypi@reimers.io" }]
license = "MIT"
readme = "README.md"
packages = [{ include = "sqlalchemy_schema" }]
dependencies = [
    "sqlalchemy >= 2.0",
    "jsonschema >= 4.23",
    "pyyaml >= 5.4",
    "loguru >= 0.7",
    "click >=8.1"
]
requires-python = ">=3.10,<4.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14"
python-dateutil = ">=2.8"
pytest = "^8.3"
pytest-cov = "^6.0"
pytest-mock = "^3.14.0"
types-pygments = "^2.18.0.20240506"
types-colorama = "^0.4.15.20240311"
types-jsonschema = "^4.23.0.20241208"
types-pytz = "^2024.2.0.20241221"
types-setuptools = "^75.6.0.20241223"
types-python-dateutil = "^2.9.0.20241206"
types-pyyaml = "^6.0.12.20241230"
types-mock = "^5.1.0.20240425"
pytest-unordered = "^0.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = ".venv/"
plugins = ["sqlalchemy.ext.mypy.plugin"]
check_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_subclassing_any = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true

[tool.pytest.ini_options]
timeout = 10

[project.scripts]
sqlalchemy_schema = "sqlalchemy_schema.command.main:main"
